Mostly formatting.

+++ src/format.cpp
@@ -3,8 +3,9 @@
 
 std::string Format::ElapsedTime(long seconds)
 {
-    int32_t hrs = seconds / 3600;
-    int32_t mins = (seconds - hrs * 3600) / 60;
-    int32_t secs = seconds - (hrs * 3600 + mins * 60);
+    long hrs = seconds / 3600;
+    long mins = (seconds - hrs * 3600) / 60;
+    long secs = seconds - (hrs * 3600 + mins * 60);
+
     return std::to_string(hrs) + ":" + std::to_string(mins) + ":" + std::to_string(secs);
 }
